user32: Always process sent messages when sending inter-thread messages.
Sometimes the message will be processed before the loop has had time to
execute once, and sent messages will remain in the queue unprocessed. This
never happens on Windows, although it is not quite clear whether this is
guaranteed or whether it is simply due to consistency in thread scheduling.
MSDN states:

 However, the sending thread will process incoming nonqueued messages while
 waiting for its message to be processed.

which is somewhat ambiguous.

We have tests in some places (namely ole32:clipboard) which were failing
intermittently because sent messages were sometimes not being processed while
executing SendMessageTimeout(). This test has never failed on Windows, in any
configuration tracked on test.winehq.org, and so I am inclined to conclude
that this logic matches Windows' behaviour.

In any case it does not seem that this should cause regressions -- any code
that was relying on SendMessage processing (or failing to process) sent
messages would have been experiencing both behaviours anyway.